🔨Refactor: 그림 맞추기 코드 구조 수정
