+++ src/app/game/(games)/drawingquiz/page.tsx
@@ -1,24 +1,5 @@
-"use client";
+import ResponsiveContainer from "../../../../containers/game/drawingquiz/ResponsiveContainer";
 
-import { useState, useEffect } from "react";
-import DesktopContainer from "../../../../containers/game/drawingquiz/DesktopContainer";
-import MobileContainer from "../../../../containers/game/drawingquiz/MobileContainer";
-
-export default function DrawingQuiz() {
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setIsMobile(window.innerWidth < 1025);
-    };
-
-    handleResize();
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  return isMobile ? <MobileContainer /> : <DesktopContainer />;
+export default function page() {
+  return <ResponsiveContainer />;
 }
